+++ src/@data/Client/clientStateLink.js
@@ -1,8 +1,8 @@
 import { withClientState } from 'apollo-link-state';
 import * as MediaPlayerQueryResolvers from '@data/mediaPlayer/queries';
 import * as MediaPlayerMutationResolvers from '@data/mediaPlayer/mutations';
-import * as GiveQueryResolvers from '@data/withGive/queries';
-import * as GiveMutationResolvers from '@data/withGive/mutations';
+import * as GiveQueryResolvers from '@data/withGive/resolvers/queries';
+import * as GiveMutationResolvers from '@data/withGive/resolvers/mutations';
 
 export default withClientState({
   Query: {
