+++ src/@data/Client/httpLink.js
@@ -1,5 +1,5 @@
 import { createHttpLink } from 'apollo-link-http';
 
 // export default createHttpLink({ uri: 'https://api.newspring.cc/graphql' });
-export default createHttpLink({ uri: 'https://alpha-api.newspring.cc/graphql' });
-// export default createHttpLink({ uri: 'http://localhost:8888/graphql' });
+// export default createHttpLink({ uri: 'https://alpha-api.newspring.cc/graphql' });
+export default createHttpLink({ uri: 'http://localhost:8888/graphql' });
