--- src/@data/withGive/isPayingMutation.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation isPaying($isPaying: Boolean!) {
-    isPaying(isPaying: $isPaying) @client
-  }
-`;
