--- src/@data/withGive/postPaymentMutation.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation postPayment {
-    postPayment @client
-  }
-`;
