--- src/@data/withGive/queries.js
@@ -1,9 +1,0 @@
-import { INITIAL_STATE } from './constants';
-
-export function contributions() {
-  return {
-    __typename: 'GiveContributions',
-    ...INITIAL_STATE,
-  };
-}
-
