+++ src/@data/withGive/resetContributions.js
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag';
+import { graphql } from 'react-apollo';
+
+export const MUTATION = gql`
+  mutation resetContributions {
+    resetContributions @client
+  }
+`;
+
+export default graphql(MUTATION, {
+  props: ({ mutate }) => ({
+    resetContributions: () => (mutate()),
+  }),
+});
