--- src/@data/withGive/resetContributionsMutation.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation resetContributions {
-    resetContributions @client
-  }
-`;
