--- src/@data/withGive/setContributionFrequencyMutation.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation setContributionFrequency($id: String!) {
-    setContributionFrequency(id: $id) @client
-  }
-`;
