--- src/@data/withGive/setContributionStartDateMutation.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation setContributionStartDate($startDate: String!) {
-    setContributionStartDate(startDate: $startDate) @client
-  }
-`;
