--- src/@data/withGive/setOrderMutation.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation setOrder($url: String!) {
-    setOrder(url: $url) @client
-  }
-`;
