--- src/@data/withGive/setPaymentMethodMutation.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation setPaymentMethod($method: String!) {
-    setPaymentMethod(method: $method) @client
-  }
-`;
