--- src/@data/withGive/validateCardMutation.js
@@ -1,11 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  mutation ValidateCard ($token: ID!) {
-    response: validate(token: $token) {
-      error
-      success
-      code
-    }
-  }
-`;
